+++ src/main/java/com/revature/Withdrawable.java
@@ -1,5 +1,5 @@
 package com.revature;
 
 public interface Withdrawable {
-	public void Withdraw(int amount);
+	public void Withdraw(double amount);
 }
